+++ test/unit/fuzz-tools.test.js
@@ -403,6 +403,7 @@ import * as marshOberhuberV423 from '../../lib/marsh-oberhuber-v423.js';
 import * as bethesdaThyroidV424 from '../../lib/bethesda-thyroid-v424.js';
 import * as vurGradeV425 from '../../lib/vur-grade-v425.js';
 import * as gellCoombsV426 from '../../lib/gell-coombs-v426.js';
+import * as vaughanWilliamsV427 from '../../lib/vaughan-williams-v427.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -751,6 +752,7 @@ const MODULES = {
   'bethesda-thyroid-v424.js': bethesdaThyroidV424,
   'vur-grade-v425.js': vurGradeV425,
   'gell-coombs-v426.js': gellCoombsV426,
+  'vaughan-williams-v427.js': vaughanWilliamsV427,
   'tb-testing.js': tbTesting,
 };